add size_t

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: asobrino <asobrino@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/10 13:05:33 by asobrino          #+#    #+#             */
-/*   Updated: 2025/04/10 13:18:43 by asobrino         ###   ########.fr       */
+/*   Updated: 2025/04/10 13:28:46 by asobrino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long unsigned int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (s[len])
